Add files via upload
this program has twoDshapes as a parent class and circle hexagon triangle and square as child classes. there are area and distance calculating functions(methods)  inside each class which can calculate the area of the shape and the distance the shape is located in comparison with the other shape on the XY plane.for more information please read the user manual.